Clean up APIClient
