+++ src/store/cache/tracks/utils/helpers.ts
@@ -1,11 +1,10 @@
-import Track from 'models/Track'
+import Track, { UserTrackMetadata } from 'models/Track'
 import User from 'models/User'
 import { Kind } from 'store/types'
 import { put } from 'redux-saga/effects'
 import * as cacheActions from 'store/cache/actions'
 import { reformat as reformatUser } from 'store/cache/users/utils'
 import { makeUid } from 'utils/uid'
-import { UserTrackMetadata } from 'services/audius-api-client/AudiusAPIClient'
 
 /**
  * Adds users from track metadata to cache.
